requirements.txt
fastapi[standard]==0.115.11
uvicorn==0.34.0
SQLAlchemy==2.0.39
databases==0.9.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
alembic==1.15.1
pydantic[email]
python-jose[cryptography]==3.4.0
passlib==1.7.4
bcrypt==4.3.0
mypy==1.15.0 